Implement gravity-type P2P emitters
While at it, improve particle spawing:

- Increase maximum number of particles to 64
- Birthrate take priority over particle lifetime

+++ src/engine/graphics/shader/shaders.h
@@ -161,7 +161,7 @@ struct ShaderParticle {
 };
 
 struct ParticlesUniforms {
-    glm::vec2 gridSize { 0.0f };
+    glm::ivec2 gridSize { 0 };
     int render { 0 };
     char padding[4];
     ShaderParticle particles[kMaxParticles];
